refactor class configuration
