+++ RussianRoulette/RussianRoulette.Api/Program.cs
@@ -1,53 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Reflection;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Converters;
-using Newtonsoft.Json.Serialization;
 
 namespace RussianRoulette.Api
 {
-    public static class Configuration
-    {
-        public static IServiceCollection ConfigureServices(this IServiceCollection services)
-        {
-            return services
-               
-                
-                .AddMvcCore()
-                .AddApiExplorer()
-                .AddNewtonsoftJson(o =>
-                {
-                    o.SerializerSettings.Converters.Add(new StringEnumConverter());
-                    o.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
-                })
-                .Services
-                .AddCors(o => o.AddPolicy("MyPolicy", builder =>
-                {
-                    builder.AllowAnyOrigin()
-                        .AllowAnyMethod()
-                        .AllowAnyHeader();
-                }))
-                .AddResponseCompression(options =>
-                {
-                    options.EnableForHttps = true;
-                    options.MimeTypes = new[] { "text/plain", "application/json" };
-                });
-
-        }
-
-        public static IApplicationBuilder Configure(IApplicationBuilder app, Func<IApplicationBuilder, IApplicationBuilder> configureHost) => configureHost(app)
-            .UseMvc(routes =>
-                routes.MapRoute("default", "{controller=Home}/{action=Index}/{id?}"));
-    }
-
     public class Program
     {
         public static void Main(string[] args)
